cleaned delete and merge code
